feat: add FK navigation, Identity/Sequence detection, and new CLI options

+++ NHTool/Helpers/TypeMapper.cs
@@ -54,8 +54,8 @@ public static class TypeMapper
 
     private static string MapOracle(ColumnInfo col) => col.DataType.ToUpperInvariant() switch
     {
-        // NUMBER without precision/scale (very common for IDs) -> int
-        "NUMBER" when col.Precision is null && col.Scale is null => "int",
+        // NUMBER without precision/scale can represent large ranges, keep it safe.
+        "NUMBER" when col.Precision is null && col.Scale is null => "decimal",
         "NUMBER" when col.Scale is null or 0 && col.Precision <= 10 => "int",
         "NUMBER" when col.Scale is null or 0 && col.Precision > 10 => "long",
         "NUMBER" => "decimal",
